requirements.txt
flask
androguard
flask-dropzone
flask-mysql
python-magic